Reducing unnecessary complexity of turbulence code: Part 1
See merge request icon/icon-nwp!1546

Avoiding the "OPTIONAL" treatment of some arrays transferred to 'turbdiff' and 'turbtran' and omitting the transfer of dispensable arrays at the initialization phase of these routines.
This changeset also removes the optional incrementation of 1-st order prognostic variables in 'turbdiff' and 'vertdiff', instead of using tendency arrays. Moreover, the unused tuning option `imode_rat_sea` is removed.

Merged-by: Guenther Zaengl <guenther.zaengl@dwd.de>

+++ src/configure_model/mo_turbdiff_config.f90
@@ -98,7 +98,7 @@ MODULE mo_turbdiff_config
 
     INTEGER:: &    ! mode to treating the aerodynamic surface-smoothing by snow 
       & imode_charpar
-    REAL(wp):: &   ! mode to treating the aerodynamic surface-smoothing by snow
+    INTEGER:: &    ! mode to treating the aerodynamic surface-smoothing by snow
       & imode_snowsmot
     REAL(wp):: &   ! lower limit of velocity-dependent Charnock-parameter
       &  alpha0    !
